+++ src/main/java/com/backend/nova/bill/controller/BillController.java
@@ -2,7 +2,6 @@ package com.backend.nova.bill.controller;
 
 import com.backend.nova.bill.dto.*;
 import com.backend.nova.bill.service.BillService;
-import com.backend.nova.common.security.SecurityUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
