+++ src/main/java/com/backend/nova/bill/service/BillService.java
@@ -6,6 +6,7 @@ import com.backend.nova.bill.dto.*;
 import com.backend.nova.bill.entity.Bill;
 import com.backend.nova.bill.entity.BillStatus;
 import com.backend.nova.bill.repository.BillRepository;
+import com.backend.nova.global.security.util.SecurityUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.Authentication;
@@ -255,7 +256,7 @@ import com.backend.nova.bill.entity.Bill;
 import com.backend.nova.bill.entity.BillStatus;
 import com.backend.nova.bill.repository.BillRepository;
 import com.backend.nova.common.exception.BusinessException;
-import com.backend.nova.common.security.SecurityUtil;
+import SecurityUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
